+++ genequery/GuiWidgets/__init__.py
@@ -1,0 +1,1 @@
+from genequery.GuiWidgets.exportDialogue import exportDialog
